Activity_Feed/ User_Profile_Page

+++ base/views.py
@@ -1,11 +1,12 @@
 from django.shortcuts import render, redirect
-#from django.http import HttpResponse
+from django.http import HttpResponse
 from django.contrib import messages #login時的error message
-#from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required #用@控制必須login
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from .models import Room,Topic
+from django.contrib.auth.forms import UserCreationForm #內建註冊表格
+from .models import Room,Topic,Message
 from .forms import RoomForm #, UserForm, MyUserCreationForm
 # Create your views here.
 
@@ -18,11 +19,11 @@ from .forms import RoomForm #, UserForm, MyUserCreationForm
 
 def loginPage(request):
     page = 'login'
-    # if request.user.is_authenticated:
-    #     return redirect('home')
+    if request.user.is_authenticated: #已登入者會直接導向home
+        return redirect('home')
 
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         email = request.POST.get('email')
         password = request.POST.get('password')
 
@@ -48,6 +49,23 @@ def logoutUser(request):
     return redirect('home')
 
 
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user) #註冊完立刻登入
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'base/login_register.html',{'form':form})
+
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -57,20 +75,47 @@ def home(request):
         )
     topics = Topic.objects.all()
     rooms_count  = rooms.count()
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains = q))
 
-    context={'rooms': rooms,'topics':topics,'rooms_count':rooms_count}
+
+    context={'rooms': rooms,
+             'topics':topics,
+             'rooms_count':rooms_count,
+             'room_messages':room_messages}
     return render (request, "base/home.html",context)
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-
+    room_messages = room.message_set.all() # query child object of a specific room:message是Model name(小寫)，跟room有關的message
+    participants = room.participants.all()
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body') #在room.html命名input為body
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk= room.id)
     context = {'room': room,
-               
+               'room_messages':room_messages,
+               'participants':participants
                }
     return render(request, "base/room.html",context)
 
 
-#@login_required(login_url='login')
+def userProfile(request,pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_message = user.message_set.all()
+    topics = Topic.objects.all()
+    context = {'user':user,
+               'rooms':rooms,
+               'room_message':room_message,
+               'topics':topics}
+    return render(request,'base/profile.html',context)
+
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     #topics = Topic.objects.all()
@@ -97,13 +142,13 @@ def createRoom(request):
 
 
 
-#@login_required(login_url='login')
+@login_required(login_url='login')
 def updateRoom(request, pk): #pk:what item are we update
     room = Room.objects.get(id=pk) #get the item we are updating
     form = RoomForm(instance=room) #表格將會prefill with room，也就是剛get到的值
     # topics = Topic.objects.all()
-    # if request.user != room.host:
-    #     return HttpResponse('Your are not allowed here!!')
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)#instance = ? :Which room to update
@@ -122,14 +167,27 @@ def updateRoom(request, pk): #pk:what item are we update
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
 
-    # if request.user != room.host:
-    #     return HttpResponse('Your are not allowed here!!')
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
 
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('Your are not allowed here!!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
